tarea-2/requirements.txt
click==8.1.7
numpy==1.25.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqtgraph==0.13.3
pyserial==3.5
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
shiboken6==6.6.0
